Better French translation for one thing
Hello !
I propose to you a better translation to the line n°38. 
In my opinion ,the correct translation for "More like this" is "Similaire". By the way this is the translation chosen in the official Emby app. 
The actual translation in French is "Plus de" ,in English that could be translated by "More of" / "More than" .